+++ ImageProcessor/src/image/images_in_memory/pgm/InMemoryPgm.java
@@ -5,7 +5,7 @@ import image.images_in_memory.InMemoryNetpbm;
 public abstract class InMemoryPgm extends InMemoryNetpbm {
     private final short[][] pixels;
 
-    public InMemoryPgm(int width, int height, int maxValue) {
+    public InMemoryPgm(int width, int height, short maxValue) {
         this.pixels = new short[width][height];
         super(width, height, maxValue);
     }
@@ -16,7 +16,7 @@ public abstract class InMemoryPgm extends InMemoryNetpbm {
     }
 
     @Override
-    public int getDefaultValue() {
+    public short getDefaultValue() {
         return 255;
     }
