+++ ImageProcessor/src/image/images_in_memory/pgm/InMemoryPgmAscii.java
@@ -4,7 +4,7 @@ import image.images_in_memory.pbm.InMemoryPbm;
 import image.signatures.FormatType;
 
 public class InMemoryPgmAscii extends InMemoryPgm {
-    public InMemoryPgmAscii(int width, int height, int maxValue) {
+    public InMemoryPgmAscii(int width, int height, short maxValue) {
         super(width, height, maxValue);
     }
