+++ ImageProcessor/src/image/parsers/binary/BinaryPgmParser.java
@@ -20,7 +20,7 @@ public class BinaryPgmParser extends NetpbmBinaryParser {
     }
 
     @Override
-    protected InMemoryNetpbm readPixels(BufferedInputStream bis, int width, int height, int maxColor) {
+    protected InMemoryNetpbm readPixels(BufferedInputStream bis, int width, int height, short maxColor) {
         InMemoryPgmBinary image = new InMemoryPgmBinary(width, height, maxColor);
 
         for (int i = 0; i < height; i++)
